# Improve readPacket performance